Check file exists or not

+++ src/frameworks/laravel/stubs/migrationStub.ts
@@ -41,6 +41,8 @@ return new class extends Migration {
     }
 };
 `;
-
+    if (fs.existsSync(filePath)) {
+        return;
+    }
     fs.writeFileSync(filePath, migration);
 }
