+++ src/frameworks/laravel/stubs/modelStub.ts
@@ -28,5 +28,9 @@ class ${module} extends Model
 }
 `;
 
+    if (fs.existsSync(filePath)) {
+        return;
+    }
+
     fs.writeFileSync(filePath, model);
 }
